requirements.txt
# ============================================================
# FILE: requirements.txt
# PASTE THIS AS: requirements.txt
# Then run: pip install -r requirements.txt
# ============================================================

# ── Web Framework ──────────────────────────────────────────
flask>=2.3.0
Werkzeug>=2.3.0

# ── Email Parsing ──────────────────────────────────────────
mail-parser>=3.15.0

# ── Data / Dataset Analysis ───────────────────────────────
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1

# ── Machine Learning ──────────────────────────────────────
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0

# ── NLP / Language Detection ──────────────────────────────
langdetect>=1.0.9

# ── Deep Learning (BERT) ──────────────────────────────────
torch>=2.0.0
transformers>=4.30.0

# ── DNS / Network ─────────────────────────────────────────
dnspython>=2.3.0
requests>=2.31.0

# ── WHOIS / Domain Intel ──────────────────────────────────
python-whois>=0.8.0

# ── PDF Report Generation ─────────────────────────────────
fpdf2>=2.7.0
reportlab>=4.0.0

# ── File Type Detection ───────────────────────────────────
python-magic>=0.4.27

# ── Explainability (SHAP) ─────────────────────────────────
shap>=0.42.0

# ── Utility ───────────────────────────────────────────────
python-dotenv>=1.0.0